perf: fp16 KV cache — halves memory bandwidth during attention
Store key/value cache in fp16 (uint16_t) instead of float32.
All models (Qwen3, Qwen3.5) affected.

Memory savings:
- Qwen3-0.6B: 28 layers x 2048 capacity x 8 kv_heads x 128 dim
  Float32: ~232 MB → Float16: ~116 MB
- Qwen3.5-0.8B: 6 full-attn layers x 2048 x 2 x 256
  Float32: ~24 MB → Float16: ~12 MB

On write: float32 cast to __fp16 (ARM64 native), or scalar f32_to_f16 fallback.
On read (gqa_attention): unwrap ring buffer + convert fp16->float32 into a
stack-allocated temporary buffer (alloca), then call gqa_attention normally.
This avoids modifying gqa_attention signature while keeping correct semantics.

For long contexts (>128 tokens), attention is memory-bandwidth bound.
fp16 KV cache reduces bandwidth by 2x for the attention read path.
Stack allocation via alloca avoids heap pressure.

+++ models/llm/qwen3.h
@@ -64,8 +64,8 @@ private:
     };
 
     struct KVCache {
-        float* k_cache = nullptr;
-        float* v_cache = nullptr;
+        uint16_t* k_cache = nullptr;  // fp16
+        uint16_t* v_cache = nullptr;  // fp16
         int len = 0;
         int start = 0;
         int capacity = 0;
